requirements.txt
python-telegram-bot==12.6.1
requests
Pillow
pymongo